Move types part of pkg exported interfaces to @0x/types from base-contract so more easily accessible
