+++ packages/base-contract/src/utils/filter_utils.ts
@@ -1,11 +1,10 @@
+import { IndexedFilterValues } from '@0x/types';
 import { BigNumber } from '@0x/utils';
 import { BlockRange, ContractAbi, EventAbi, FilterObject, LogEntry } from 'ethereum-types';
 import * as ethUtil from 'ethereumjs-util';
 import * as jsSHA3 from 'js-sha3';
 import * as _ from 'lodash';
 import * as uuid from 'uuid/v4';
-
-import { IndexedFilterValues } from '../types';
 
 const TOPIC_LENGTH = 32;
