+++ packages/types/src/index.ts
@@ -1,7 +1,7 @@
 // tslint:disable:max-file-line-count
 
 import { BigNumber } from 'bignumber.js';
-import { ContractAbi, ContractNetworks, DevdocOutput } from 'ethereum-types';
+import { ContractAbi, ContractEventArg, ContractNetworks, DecodedLogArgs, DevdocOutput, LogWithDecodedArgs } from 'ethereum-types';
 
 export interface Order {
     chainId: number;
@@ -840,4 +840,18 @@ export interface OrderInfo {
     orderStatus: number;
     orderHash: string;
     orderTakerAssetFilledAmount: BigNumber;
+}
+
+export interface DecodedLogEvent<ArgsType extends DecodedLogArgs> {
+    isRemoved: boolean;
+    log: LogWithDecodedArgs<ArgsType>;
+}
+
+export type EventCallback<ArgsType extends DecodedLogArgs> = (
+    err: null | Error,
+    log?: DecodedLogEvent<ArgsType>,
+) => void;
+
+export interface IndexedFilterValues {
+    [index: string]: ContractEventArg;
 }
